long

+++ manydepth/trainer_m2_normals.py
@@ -766,7 +766,7 @@ class Trainer_Monodepth2:
         bottom_right_depth = bottom_right_flat.permute(0, 2, 1).to(device=K_inv.device) * D.view(batch_size, 1, -1)
         top_right_depth = top_right_flat.permute(0, 2, 1).to(device=K_inv.device) * D.view(batch_size, 1, -1)
         bottom_left_depth = bottom_left_flat.permute(0, 2, 1).to(device=K_inv.device) * D.view(batch_size, 1, -1)
-        top_left_depth = torch.gather(D, 2, top_left_flat.unsqueeze(1))
+        top_left_depth = torch.gather(D, 2, top_left_flat.unsqueeze(1).long())
         print(top_left_depth.shape)
         #print(ones.shape)
